Add useInitFirebase hook for Firebase messaging and fix imports in user page
